webstorage/localstorage-about-blank-3P-window-partitioned.tentative.sub.html

<!doctype html>
<meta charset=utf-8>
<title>localStorage: about:blank partitioning</title>
<meta name=help href="https://privacycg.github.io/storage-partitioning/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/webstorage/resources/partitioning-utils.js"></script>
<style>iframe {display: none}</style>
<body>
<script>
promise_test(async t => {
  localStorage.clear();

  const path =
    "webstorage/resources/localstorage-about-blank-partitioned-win-open.html";
  const crossSiteURL = `http://{{hosts[alt][]}}:{{ports[http][0]}}/${path}`;
  const sameSiteURL = `http://{{hosts[][]}}:{{ports[http][0]}}/${path}`;
  let firstPartyID = getOrCreateID("userID4");
  let crossSiteWinID;
  let sameSiteWinID;
  let crossSiteWin;
  let sameSiteWin;
  let crossSiteAboutBlankID;
  let sameSiteAboutBlankID;

  window.addEventListener("message", e => {
    const payload = {
      command: "open about:blank window"
    }

    if (e.data.message === "window loaded") {
      if (crossSiteURL.includes(e.origin)) {
        // Step 2. cross-site window is loaded, capture reference to its ID
        crossSiteWinID = e.data.userID;
        // Step 3. now, create a same-site iframe
        sameSiteWin = window.open(sameSiteURL, "", "noopener=false");
        // Step 4. Ask the cross-site window to create an about:blank window
        crossSiteWin.postMessage(payload, e.origin);
      }

      if (sameSiteURL.includes(e.origin)) {
        // Step 5. same-site window is loaded, capture reference to its ID
        sameSiteWinID = e.data.userID;
        // Step 6. Ask the same-site window to create an about:blank window
        sameSiteWin.postMessage(payload, e.origin);
      }
    }

    if (e.data.message === "about:blank frame ID") {
      // Step 7. capture reference to same-site window ID
      if (crossSiteURL.includes(e.origin)) {
        crossSiteAboutBlankID = e.data.userID;
      }

      // Step 8. capture reference to same-site window ID
      if (sameSiteURL.includes(e.origin)) {
        sameSiteAboutBlankID = e.data.userID;

        // Step 9. Assert some things
        assert_true(sameSiteAboutBlankID !== undefined);
        assert_true(crossSiteAboutBlankID !== undefined);
        assert_true(sameSiteWinID === sameSiteAboutBlankID);
        assert_true(crossSiteWinID === crossSiteAboutBlankID);
        assert_false(sameSiteAboutBlankID === crossSiteAboutBlankID);
        sameSiteWin.postMessage({command: "close about:blank window"}, "*");
        crossSiteWin.postMessage({command: "close about:blank window"}, "*");
        // create a delay so we don't race with the about:blank close
        setTimeout(() => {
          sameSiteWin.close();
          crossSiteWin.close();
        }, 250);
      }
    }
  });

  // Step 1. Add a cross-site iframe
  crossSiteWin = window.open(crossSiteURL, "", "noopener=false");
}, "about:blank win opened from 3P window.open not partitioned with 1P frame.");

</script>
</body>
